move type declarations from interface/Product to lib/backend
