+++ e-commerce/shop-frontend/app/checkout/page.tsx
@@ -1,8 +1,7 @@
 import Checkout from './checkout-client'
-import { CartProduct } from '@/interfaces/Product';
 import { cookies } from "next/headers";
 import { getRequestCookie } from "@/lib/session";
-import { api } from '@/lib/backend';
+import { api, CartProduct } from '@/lib/backend';
 import { redirect } from "next/navigation";
 import { HttpError, ok } from 'oazapfts';
